Revert move to Lib
